test(KEN-1241): split session-lanes by surface; one lifecycle table
pi-agents-tmux's session-lanes.test.ts held five surfaces in 2,439 lines:
lane assignment, the one-shot transcript, the one-shot process lifecycle,
completion classification and the dispatch concurrency helper. Each moves
whole into a file named for it, over one shared fixture module (the neutral
world: temp runtimes, the settings writer, the two spawn mocks, the bridge
event shapes).

The fifteen `bg one-shot` cases that staged one lifecycle each through the
same mock and asserted the same outcome fields become one table with one
deepEqual per row over the observed outcome. Six cases with their own
apparatus (a throwing update callback, a process.kill override, the
original-deadline timing, an abort, a real child process) and the existing
three-row exit table stay as they were. Prose pins on errorMessage survive
only where the message is the sole carrier of the fact; every other outcome
is read from the exit code, stop reason, kill sequence, lifecycle event or
a transcript record type.

Mutation: reporting a settled skip as newer_agent_generation instead of
agent_active fails the stale-settlement row.
